Moved window javascript event in JSFunction
