+++ HTML_Renderer/src/main/java/org/lobobrowser/html/js/Window.java
@@ -18,9 +18,6 @@
 
     Contact info: lobochief@users.sourceforge.net; ivan.difrancesco@yahoo.it
  */
-/*
- * Created on Nov 12, 2005
- */
 package org.lobobrowser.html.js;
 
 import java.awt.event.ActionEvent;
@@ -29,7 +26,6 @@ import java.lang.ref.Reference;
 import java.lang.ref.WeakReference;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.WeakHashMap;
 
@@ -41,22 +37,7 @@ import org.apache.logging.log4j.Logger;
 import org.lobobrowser.html.HtmlRendererContext;
 import org.lobobrowser.html.domimpl.HTMLDocumentImpl;
 import org.lobobrowser.html.domimpl.HTMLElementImpl;
-import org.lobobrowser.html.domimpl.HTMLIFrameElementImpl;
-import org.lobobrowser.html.domimpl.HTMLImageElementImpl;
-import org.lobobrowser.html.domimpl.HTMLOptionElementImpl;
-import org.lobobrowser.html.domimpl.HTMLScriptElementImpl;
-import org.lobobrowser.html.domimpl.HTMLSelectElementImpl;
-import org.lobobrowser.html.js.xml.XMLHttpRequest;
-import org.lobobrowser.html.js.xml.XMLSerializer;
-import org.lobobrowser.html.jsimpl.ConsoleImpl;
-import org.lobobrowser.html.xpath.XPathResultImpl;
 import org.lobobrowser.http.UserAgentContext;
-import org.lobobrowser.js.AbstractScriptableDelegate;
-import org.lobobrowser.js.JavaClassWrapper;
-import org.lobobrowser.js.JavaClassWrapperFactory;
-import org.lobobrowser.js.JavaInstantiator;
-import org.lobobrowser.js.JavaObjectWrapper;
-import org.lobobrowser.js.JavaScript;
 import org.lobobrowser.util.ID;
 import org.lobobrowser.w3c.html.ApplicationCache;
 import org.lobobrowser.w3c.html.BarProp;
@@ -69,7 +50,6 @@ import org.lobobrowser.w3c.html.UndoManager;
 import org.mozilla.javascript.Context;
 import org.mozilla.javascript.Function;
 import org.mozilla.javascript.Scriptable;
-import org.mozilla.javascript.ScriptableObject;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
@@ -80,25 +60,13 @@ import org.w3c.dom.views.DocumentView;
 /**
  * The Class Window.
  */
-public class Window extends AbstractScriptableDelegate implements AbstractView {
+public class Window extends JSFunction implements AbstractView {
 
 	/** The Constant logger. */
 	private static final Logger logger = LogManager.getLogger(Window.class.getName());
 
 	/** The Constant CONTEXT_WINDOWS. */
 	private static final Map<HtmlRendererContext, WeakReference<Window>> CONTEXT_WINDOWS = new WeakHashMap<HtmlRendererContext, WeakReference<Window>>();
-
-	/** The Constant XMLHTTPREQUEST_WRAPPER. */
-	private static final JavaClassWrapper XMLHTTPREQUEST_WRAPPER = JavaClassWrapperFactory.getInstance()
-			.getClassWrapper(XMLHttpRequest.class);
-
-	/** The Constant DOMPARSER_WRAPPER. */
-	private static final JavaClassWrapper DOMPARSER_WRAPPER = JavaClassWrapperFactory.getInstance()
-			.getClassWrapper(DOMParser.class);
-
-	/** The Constant XMLSERIALIZER_WRAPPER. */
-	private static final JavaClassWrapper XMLSERIALIZER_WRAPPER = JavaClassWrapperFactory.getInstance()
-			.getClassWrapper(XMLSerializer.class);
 
 	/** The timer id counter. */
 	private static int timerIdCounter = 0;
@@ -118,20 +86,8 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	/** The location. */
 	private Location location;
 
-	/** The window scope. */
-	private ScriptableObject windowScope;
-
 	/** The history. */
 	private History history;
-
-	/** The onunload. */
-	private Function onunload;
-
-	/** The task map. */
-	private Map<Integer, TaskWrapper> taskMap;
-
-	/** The document. */
-	private volatile HTMLDocumentImpl document;
 
 	/** The length. */
 	private int length;
@@ -144,184 +100,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 
 	/** The name set. */
 	private boolean nameSet = false;
-
-	/** The onabort. */
-	private Function onabort;
-
-	/** The onafterprint. */
-	private Function onafterprint;
-
-	/** The onbeforeprint. */
-	private Function onbeforeprint;
-
-	/** The onbeforeunload. */
-	private Function onbeforeunload;
-
-	/** The onblur. */
-	private Function onblur;
-
-	/** The oncanplay. */
-	private Function oncanplay;
-
-	/** The oncanplaythrough. */
-	private Function oncanplaythrough;
-
-	/** The onchange. */
-	private Function onchange;
-
-	/** The oncontextmenu. */
-	private Function oncontextmenu;
-
-	/** The ondrag. */
-	private Function ondrag;
-
-	/** The ondragend. */
-	private Function ondragend;
-
-	/** The ondragenter. */
-	private Function ondragenter;
-
-	/** The ondragleave. */
-	private Function ondragleave;
-
-	/** The ondragover. */
-	private Function ondragover;
-
-	/** The ondragstart. */
-	private Function ondragstart;
-
-	/** The ondrop. */
-	private Function ondrop;
-
-	/** The ondurationchange. */
-	private Function ondurationchange;
-
-	/** The onemptied. */
-	private Function onemptied;
-
-	/** The onended. */
-	private Function onended;
-
-	/** The onerror. */
-	private Function onerror;
-
-	/** The onfocus. */
-	private Function onfocus;
-
-	/** The onformchange. */
-	private Function onformchange;
-
-	/** The onforminput. */
-	private Function onforminput;
-
-	/** The onhashchange. */
-	private Function onhashchange;
-
-	/** The oninput. */
-	private Function oninput;
-
-	/** The oninvalid. */
-	private Function oninvalid;
-
-	/** The onloadeddata. */
-	private Function onloadeddata;
-
-	/** The onloadedmetadata. */
-	private Function onloadedmetadata;
-
-	/** The onloadstart. */
-	private Function onloadstart;
-
-	/** The onmessage. */
-	private Function onmessage;
-
-	/** The onmousemove. */
-	private Function onmousemove;
-
-	/** The onmouseout. */
-	private Function onmouseout;
-
-	/** The onmousewheel. */
-	private Function onmousewheel;
-
-	/** The onoffline. */
-	private Function onoffline;
-
-	/** The ononline. */
-	private Function ononline;
-
-	/** The onpause. */
-	private Function onpause;
-
-	/** The onplay. */
-	private Function onplay;
-
-	/** The onplaying. */
-	private Function onplaying;
-
-	/** The onpagehide. */
-	private Function onpagehide;
-
-	/** The onpageshow. */
-	private Function onpageshow;
-
-	/** The onpopstate. */
-	private Function onpopstate;
-
-	/** The onprogress. */
-	private Function onprogress;
-
-	/** The onratechange. */
-	private Function onratechange;
-
-	/** The onreadystatechange. */
-	private Function onreadystatechange;
-
-	/** The onredo. */
-	private Function onredo;
-
-	/** The onresize. */
-	private Function onresize;
-
-	/** The onscroll. */
-	private Function onscroll;
-
-	/** The onseeked. */
-	private Function onseeked;
-
-	/** The onseeking. */
-	private Function onseeking;
-
-	/** The onselect. */
-	private Function onselect;
-
-	/** The onshow. */
-	private Function onshow;
-
-	/** The onstalled. */
-	private Function onstalled;
-
-	/** The onstorage. */
-	private Function onstorage;
-
-	/** The onsubmit. */
-	private Function onsubmit;
-
-	/** The onsuspend. */
-	private Function onsuspend;
-
-	/** The ontimeupdate. */
-	private Function ontimeupdate;
-
-	/** The onundo. */
-	private Function onundo;
-
-	/** The onvolumechange. */
-	private Function onvolumechange;
-
-	/** The onwaiting. */
-	private Function onwaiting;
-
+	
 	/**
 	 * Instantiates a new window.
 	 *
@@ -331,13 +110,14 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 *            the ua context
 	 */
 	public Window(HtmlRendererContext rcontext, UserAgentContext uaContext) {
+		super(uaContext);
 		// TODO: Probably need to create a new Window instance
 		// for every document. Sharing of Window state between
 		// different documents is not correct.
 		this.rcontext = rcontext;
 		this.uaContext = uaContext;
 	}
-
+	
 	/**
 	 * Generate timer id.
 	 *
@@ -367,57 +147,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 		return this.uaContext;
 	}
 
-	/**
-	 * Clear state.
-	 */
-	private void clearState() {
-		Context.enter();
-		try {
-			Scriptable s = this.getWindowScope();
-			if (s != null) {
-				Object[] ids = s.getIds();
-				for (Object id : ids) {
-					if (id instanceof String) {
-						s.delete((String) id);
-					} else if (id instanceof Integer) {
-						s.delete(((Integer) id).intValue());
-					}
-				}
-			}
-		} finally {
-			Context.exit();
-		}
-
-	}
-
-	/**
-	 * Sets the document.
-	 *
-	 * @param document
-	 *            the new document
-	 */
-	public void setDocument(HTMLDocumentImpl document) {
-		Document prevDocument = this.document;
-		if (prevDocument != document) {
-			// Should clearing of the state be done
-			// when window "unloads"?
-			if (prevDocument != null) {
-				// Only clearing when the previous document was not null
-				// because state might have been set on the window before
-				// the very first document is added.
-				this.clearState();
-			}
-			this.initWindowScope(document);
-			this.forgetAllTasks();
-			Function onunload = this.onunload;
-			if (onunload != null) {
-				HTMLDocumentImpl oldDoc = this.document;
-				Executor.executeFunction(this.getWindowScope(), onunload, oldDoc.getDocumentURL(), this.uaContext);
-				this.onunload = null;
-			}
-			this.document = document;
-		}
-	}
+	
 
 	/*
 	 * (non-Javadoc)
@@ -426,7 +156,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 */
 	@Override
 	public DocumentView getDocument() {
-		return this.document;
+		return this.getWindowDocument();
 	}
 
 	/**
@@ -435,76 +165,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 * @return the document node
 	 */
 	public Document getDocumentNode() {
-		return this.document;
-	}
-
-	/**
-	 * Put and start task.
-	 *
-	 * @param timeoutID
-	 *            the timeout id
-	 * @param timer
-	 *            the timer
-	 * @param retained
-	 *            the retained
-	 */
-	private void putAndStartTask(Integer timeoutID, Timer timer, Object retained) {
-		TaskWrapper oldTaskWrapper = null;
-		synchronized (this) {
-			Map<Integer, TaskWrapper> taskMap = this.taskMap;
-			if (taskMap == null) {
-				taskMap = new HashMap<Integer, TaskWrapper>(4);
-				this.taskMap = taskMap;
-			} else {
-				oldTaskWrapper = taskMap.get(timeoutID);
-			}
-			taskMap.put(timeoutID, new TaskWrapper(timer, retained));
-		}
-		// Do this outside synchronized block, just in case.
-		if (oldTaskWrapper != null) {
-			oldTaskWrapper.timer.stop();
-		}
-		timer.start();
-	}
-
-	/**
-	 * Forget task.
-	 *
-	 * @param timeoutID
-	 *            the timeout id
-	 * @param cancel
-	 *            the cancel
-	 */
-	private void forgetTask(Integer timeoutID, boolean cancel) {
-		TaskWrapper oldTimer = null;
-		synchronized (this) {
-			Map<Integer, TaskWrapper> taskMap = this.taskMap;
-			if (taskMap != null) {
-				oldTimer = taskMap.remove(timeoutID);
-			}
-		}
-		if (oldTimer != null && cancel) {
-			oldTimer.timer.stop();
-		}
-	}
-
-	/**
-	 * Forget all tasks.
-	 */
-	private void forgetAllTasks() {
-		TaskWrapper[] oldTaskWrappers = null;
-		synchronized (this) {
-			Map<Integer, TaskWrapper> taskMap = this.taskMap;
-			if (taskMap != null) {
-				oldTaskWrappers = taskMap.values().toArray(new TaskWrapper[0]);
-				this.taskMap = null;
-			}
-		}
-		if (oldTaskWrappers != null) {
-			for (TaskWrapper taskWrapper : oldTaskWrappers) {
-				taskWrapper.timer.stop();
-			}
-		}
+		return this.getWindowDocument();
 	}
 
 	/**
@@ -658,7 +319,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 * @return the object
 	 */
 	public Object eval(String javascript) {
-		HTMLDocumentImpl document = this.document;
+		HTMLDocumentImpl document = this.getWindowDocument();
 		if (document == null) {
 			throw new IllegalStateException("Cannot evaluate if document is not set.");
 		}
@@ -689,109 +350,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 			rcontext.focus();
 		}
 	}
-
-	/**
-	 * Inits the window scope.
-	 *
-	 * @param doc
-	 *            the doc
-	 */
-	private void initWindowScope(final Document doc) {
-		final Scriptable ws = this.getWindowScope();
-		JavaInstantiator jiXhttp = () -> {
-			Document d = doc;
-			if (d == null) {
-				throw new IllegalStateException("Cannot perform operation when document is unset.");
-			}
-			HTMLDocumentImpl hd;
-			try {
-				hd = (HTMLDocumentImpl) d;
-			} catch (ClassCastException err) {
-				throw new IllegalStateException(
-						"Cannot perform operation with documents of type " + d.getClass().getName() + ".");
-			}
-			return new XMLHttpRequest(uaContext, hd.getDocumentURL(), ws);
-		};
-
-		JavaInstantiator jiDomParser = () -> new DOMParser();
-
-		JavaInstantiator jiXMLSerializer = () -> new XMLSerializer();
-
-		Function xmlHttpRequestC = JavaObjectWrapper.getConstructor("XMLHttpRequest", XMLHTTPREQUEST_WRAPPER, ws,
-				jiXhttp);
-		ScriptableObject.defineProperty(ws, "XMLHttpRequest", xmlHttpRequestC, ScriptableObject.READONLY);
-
-		Function domParser = JavaObjectWrapper.getConstructor("DOMParser", DOMPARSER_WRAPPER, ws, jiDomParser);
-		ScriptableObject.defineProperty(ws, "DOMParser", domParser, ScriptableObject.READONLY);
-
-		Function xmlserial = JavaObjectWrapper.getConstructor("XMLSerializer", XMLSERIALIZER_WRAPPER, ws,
-				jiXMLSerializer);
-		ScriptableObject.defineProperty(ws, "XMLSerializer", xmlserial, ScriptableObject.READONLY);
-
-		// HTML element classes
-		this.defineElementClass(ws, doc, "Image", "img", HTMLImageElementImpl.class);
-		this.defineElementClass(ws, doc, "Script", "script", HTMLScriptElementImpl.class);
-		this.defineElementClass(ws, doc, "IFrame", "iframe", HTMLIFrameElementImpl.class);
-		this.defineElementClass(ws, doc, "Option", "option", HTMLOptionElementImpl.class);
-		this.defineElementClass(ws, doc, "Select", "select", HTMLSelectElementImpl.class);
-	}
-
-	/**
-	 * Gets the window scope.
-	 *
-	 * @return the window scope
-	 */
-	public Scriptable getWindowScope() {
-		synchronized (this) {
-			ScriptableObject windowScope = this.windowScope;
-			if (windowScope != null) {
-				return windowScope;
-			}
-			try {
-				Context ctx = Context.enter();
-				// Window scope needs to be top-most scope.
-				windowScope = (ScriptableObject) JavaScript.getInstance().getJavascriptObject(this, null);
-				ctx.initStandardObjects(windowScope);
-				Object consoleJSObj = JavaScript.getInstance().getJavascriptObject(new ConsoleImpl(), windowScope);
-				ScriptableObject.putProperty(windowScope, "console", consoleJSObj);
-				Object xpathresult = JavaScript.getInstance().getJavascriptObject(new XPathResultImpl(), windowScope);
-				ScriptableObject.putProperty(windowScope, "XPathResult", xpathresult);
-				this.windowScope = windowScope;
-				return windowScope;
-			} finally {
-				Context.exit();
-			}
-		}
-	}
-
-	/**
-	 * Define element class.
-	 *
-	 * @param scope
-	 *            the scope
-	 * @param document
-	 *            the document
-	 * @param jsClassName
-	 *            the js class name
-	 * @param elementName
-	 *            the element name
-	 * @param javaClass
-	 *            the java class
-	 */
-	private final void defineElementClass(Scriptable scope, final Document document, final String jsClassName,
-			final String elementName, Class<?> javaClass) {
-		JavaInstantiator ji = () -> {
-			Document d = document;
-			if (d == null) {
-				throw new IllegalStateException("Document not set in current context.");
-			}
-			return d.createElement(elementName);
-		};
-		JavaClassWrapper classWrapper = JavaClassWrapperFactory.getInstance().getClassWrapper(javaClass);
-		Function constructorFunction = JavaObjectWrapper.getConstructor(jsClassName, classWrapper, scope, ji);
-		ScriptableObject.defineProperty(scope, jsClassName, constructorFunction, ScriptableObject.READONLY);
-	}
-
+	
 	/**
 	 * Gets the window.
 	 *
@@ -834,7 +393,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 		HtmlRendererContext rcontext = this.rcontext;
 		if (rcontext != null) {
 			URL url;
-			Object document = this.document;
+			Object document = this.getWindowDocument();
 			if (document instanceof HTMLDocumentImpl) {
 				url = ((HTMLDocumentImpl) document).getFullURL(relativeUrl);
 			} else {
@@ -1090,7 +649,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 * @return the frames
 	 */
 	public HTMLCollection getFrames() {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			return ((HTMLDocumentImpl) doc).getFrames();
 		}
@@ -1363,7 +922,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 * @return the onload
 	 */
 	public Function getOnload() {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			return ((HTMLDocumentImpl) doc).getOnloadHandler();
 		} else {
@@ -1380,29 +939,10 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	public void setOnload(Function onload) {
 		// Note that body.onload overrides
 		// window.onload.
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			((HTMLDocumentImpl) doc).setOnloadHandler(onload);
 		}
-	}
-
-	/**
-	 * Gets the onunload.
-	 *
-	 * @return the onunload
-	 */
-	public Function getOnunload() {
-		return onunload;
-	}
-
-	/**
-	 * Sets the onunload.
-	 *
-	 * @param onunload
-	 *            the new onunload
-	 */
-	public void setOnunload(Function onunload) {
-		this.onunload = onunload;
 	}
 
 	/**
@@ -1413,7 +953,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 * @return the node
 	 */
 	public Node namedItem(String name) {
-		HTMLDocumentImpl doc = this.document;
+		HTMLDocumentImpl doc = this.getWindowDocument();
 		if (doc == null) {
 			return null;
 		}
@@ -1598,28 +1138,6 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	}
 
 	/**
-	 * The Class TaskWrapper.
-	 */
-	private static class TaskWrapper {
-
-		/** The timer. */
-		public final Timer timer;
-
-		/**
-		 * Instantiates a new task wrapper.
-		 *
-		 * @param timer
-		 *            the timer
-		 * @param retained
-		 *            the retained
-		 */
-		public TaskWrapper(Timer timer, Object retained) {
-			super();
-			this.timer = timer;
-		}
-	}
-
-	/**
 	 * Adds the event listener.
 	 *
 	 * @param script
@@ -1628,14 +1146,14 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 *            the function
 	 */
 	public void addEventListener(String script, Function function) {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			((HTMLDocumentImpl) doc).addEventListener(script, function);
 		}
 	}
 
 	public void removeEventListener(String script, Function function) {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			((HTMLDocumentImpl) doc).removeEventListener(script, function);
 		}
@@ -1889,165 +1407,13 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	}
 
 	/**
-	 * Gets the onabort.
-	 *
-	 * @return the onabort
-	 */
-	public Function getOnabort() {
-		return onabort;
-	}
-
-	/**
-	 * Sets the onabort.
-	 *
-	 * @param onabort
-	 *            the new onabort
-	 */
-	public void setOnabort(Function onabort) {
-		this.onabort = onabort;
-	}
-
-	/**
-	 * Gets the onafterprint.
-	 *
-	 * @return the onafterprint
-	 */
-	public Function getOnafterprint() {
-		return onafterprint;
-	}
-
-	/**
-	 * Sets the onafterprint.
-	 *
-	 * @param onafterprint
-	 *            the new onafterprint
-	 */
-	public void setOnafterprint(Function onafterprint) {
-		this.onafterprint = onafterprint;
-	}
-
-	/**
-	 * Gets the onbeforeprint.
-	 *
-	 * @return the onbeforeprint
-	 */
-	public Function getOnbeforeprint() {
-		return onbeforeprint;
-	}
-
-	/**
-	 * Sets the onbeforeprint.
-	 *
-	 * @param onbeforeprint
-	 *            the new onbeforeprint
-	 */
-	public void setOnbeforeprint(Function onbeforeprint) {
-		this.onbeforeprint = onbeforeprint;
-	}
-
-	/**
-	 * Gets the onbeforeunload.
-	 *
-	 * @return the onbeforeunload
-	 */
-	public Function getOnbeforeunload() {
-		return onbeforeunload;
-	}
-
-	/**
-	 * Sets the onbeforeunload.
-	 *
-	 * @param onbeforeunload
-	 *            the new onbeforeunload
-	 */
-	public void setOnbeforeunload(Function onbeforeunload) {
-		this.onbeforeunload = onbeforeunload;
-	}
-
-	/**
-	 * Gets the onblur.
-	 *
-	 * @return the onblur
-	 */
-	public Function getOnblur() {
-		return onblur;
-	}
-
-	/**
-	 * Sets the onblur.
-	 *
-	 * @param onblur
-	 *            the new onblur
-	 */
-	public void setOnblur(Function onblur) {
-		this.onblur = onblur;
-	}
-
-	/**
-	 * Gets the oncanplay.
-	 *
-	 * @return the oncanplay
-	 */
-	public Function getOncanplay() {
-		return oncanplay;
-	}
-
-	/**
-	 * Sets the oncanplay.
-	 *
-	 * @param oncanplay
-	 *            the new oncanplay
-	 */
-	public void setOncanplay(Function oncanplay) {
-		this.oncanplay = oncanplay;
-	}
-
-	/**
-	 * Gets the oncanplaythrough.
-	 *
-	 * @return the oncanplaythrough
-	 */
-	public Function getOncanplaythrough() {
-		return oncanplaythrough;
-	}
-
-	/**
-	 * Sets the oncanplaythrough.
-	 *
-	 * @param oncanplaythrough
-	 *            the new oncanplaythrough
-	 */
-	public void setOncanplaythrough(Function oncanplaythrough) {
-		this.oncanplaythrough = oncanplaythrough;
-	}
-
-	/**
-	 * Gets the onchange.
-	 *
-	 * @return the onchange
-	 */
-	public Function getOnchange() {
-		return onchange;
-	}
-
-	/**
-	 * Sets the onchange.
-	 *
-	 * @param onchange
-	 *            the new onchange
-	 */
-	public void setOnchange(Function onchange) {
-		this.onchange = onchange;
-	}
-
-	/**
 	 * Gets the onclick.
 	 *
 	 * @return the onclick
 	 */
 	public Function getOnclick() {
 
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			return ((HTMLDocumentImpl) doc).getOnclick();
 		} else {
@@ -2062,30 +1428,11 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 *            the new onclick
 	 */
 	public void setOnclick(Function onclick) {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			((HTMLDocumentImpl) doc).setOnclick(onclick);
 
 		}
-	}
-
-	/**
-	 * Gets the oncontextmenu.
-	 *
-	 * @return the oncontextmenu
-	 */
-	public Function getOncontextmenu() {
-		return oncontextmenu;
-	}
-
-	/**
-	 * Sets the oncontextmenu.
-	 *
-	 * @param oncontextmenu
-	 *            the new oncontextmenu
-	 */
-	public void setOncontextmenu(Function oncontextmenu) {
-		this.oncontextmenu = oncontextmenu;
 	}
 
 	/**
@@ -2094,7 +1441,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 * @return the ondblclick
 	 */
 	public Function getOndblclick() {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			return ((HTMLDocumentImpl) doc).getOndblclick();
 		} else {
@@ -2109,334 +1456,11 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 *            the new ondblclick
 	 */
 	public void setOndblclick(Function ondblclick) {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			((HTMLDocumentImpl) doc).setOndblclick(ondblclick);
 
 		}
-	}
-
-	/**
-	 * Gets the ondrag.
-	 *
-	 * @return the ondrag
-	 */
-	public Function getOndrag() {
-		return ondrag;
-	}
-
-	/**
-	 * Sets the ondrag.
-	 *
-	 * @param ondrag
-	 *            the new ondrag
-	 */
-	public void setOndrag(Function ondrag) {
-		this.ondrag = ondrag;
-	}
-
-	/**
-	 * Gets the ondragend.
-	 *
-	 * @return the ondragend
-	 */
-	public Function getOndragend() {
-		return ondragend;
-	}
-
-	/**
-	 * Sets the ondragend.
-	 *
-	 * @param ondragend
-	 *            the new ondragend
-	 */
-	public void setOndragend(Function ondragend) {
-		this.ondragend = ondragend;
-	}
-
-	/**
-	 * Gets the ondragenter.
-	 *
-	 * @return the ondragenter
-	 */
-	public Function getOndragenter() {
-		return ondragenter;
-	}
-
-	/**
-	 * Sets the ondragenter.
-	 *
-	 * @param ondragenter
-	 *            the new ondragenter
-	 */
-	public void setOndragenter(Function ondragenter) {
-		this.ondragenter = ondragenter;
-	}
-
-	/**
-	 * Gets the ondragleave.
-	 *
-	 * @return the ondragleave
-	 */
-	public Function getOndragleave() {
-		return ondragleave;
-	}
-
-	/**
-	 * Sets the ondragleave.
-	 *
-	 * @param ondragleave
-	 *            the new ondragleave
-	 */
-	public void setOndragleave(Function ondragleave) {
-		this.ondragleave = ondragleave;
-	}
-
-	/**
-	 * Gets the ondragover.
-	 *
-	 * @return the ondragover
-	 */
-	public Function getOndragover() {
-		return ondragover;
-	}
-
-	/**
-	 * Sets the ondragover.
-	 *
-	 * @param ondragover
-	 *            the new ondragover
-	 */
-	public void setOndragover(Function ondragover) {
-		this.ondragover = ondragover;
-	}
-
-	/**
-	 * Gets the ondragstart.
-	 *
-	 * @return the ondragstart
-	 */
-	public Function getOndragstart() {
-		return ondragstart;
-	}
-
-	/**
-	 * Sets the ondragstart.
-	 *
-	 * @param ondragstart
-	 *            the new ondragstart
-	 */
-	public void setOndragstart(Function ondragstart) {
-		this.ondragstart = ondragstart;
-	}
-
-	/**
-	 * Gets the ondrop.
-	 *
-	 * @return the ondrop
-	 */
-	public Function getOndrop() {
-		return ondrop;
-	}
-
-	/**
-	 * Sets the ondrop.
-	 *
-	 * @param ondrop
-	 *            the new ondrop
-	 */
-	public void setOndrop(Function ondrop) {
-		this.ondrop = ondrop;
-	}
-
-	/**
-	 * Gets the ondurationchange.
-	 *
-	 * @return the ondurationchange
-	 */
-	public Function getOndurationchange() {
-		return ondurationchange;
-	}
-
-	/**
-	 * Sets the ondurationchange.
-	 *
-	 * @param ondurationchange
-	 *            the new ondurationchange
-	 */
-	public void setOndurationchange(Function ondurationchange) {
-		this.ondurationchange = ondurationchange;
-	}
-
-	/**
-	 * Gets the onemptied.
-	 *
-	 * @return the onemptied
-	 */
-	public Function getOnemptied() {
-		return onemptied;
-	}
-
-	/**
-	 * Sets the onemptied.
-	 *
-	 * @param onemptied
-	 *            the new onemptied
-	 */
-	public void setOnemptied(Function onemptied) {
-		this.onemptied = onemptied;
-	}
-
-	/**
-	 * Gets the onended.
-	 *
-	 * @return the onended
-	 */
-	public Function getOnended() {
-		return onended;
-	}
-
-	/**
-	 * Sets the onended.
-	 *
-	 * @param onended
-	 *            the new onended
-	 */
-	public void setOnended(Function onended) {
-		this.onended = onended;
-	}
-
-	/**
-	 * Gets the onerror.
-	 *
-	 * @return the onerror
-	 */
-	public Function getOnerror() {
-		return onerror;
-	}
-
-	/**
-	 * Sets the onerror.
-	 *
-	 * @param onerror
-	 *            the new onerror
-	 */
-	public void setOnerror(Function onerror) {
-		this.onerror = onerror;
-	}
-
-	/**
-	 * Gets the onfocus.
-	 *
-	 * @return the onfocus
-	 */
-	public Function getOnfocus() {
-		return onfocus;
-	}
-
-	/**
-	 * Sets the onfocus.
-	 *
-	 * @param onfocus
-	 *            the new onfocus
-	 */
-	public void setOnfocus(Function onfocus) {
-		this.onfocus = onfocus;
-	}
-
-	/**
-	 * Gets the onformchange.
-	 *
-	 * @return the onformchange
-	 */
-	public Function getOnformchange() {
-		return onformchange;
-	}
-
-	/**
-	 * Sets the onformchange.
-	 *
-	 * @param onformchange
-	 *            the new onformchange
-	 */
-	public void setOnformchange(Function onformchange) {
-		this.onformchange = onformchange;
-	}
-
-	/**
-	 * Gets the onforminput.
-	 *
-	 * @return the onforminput
-	 */
-	public Function getOnforminput() {
-		return onforminput;
-	}
-
-	/**
-	 * Sets the onforminput.
-	 *
-	 * @param onforminput
-	 *            the new onforminput
-	 */
-	public void setOnforminput(Function onforminput) {
-		this.onforminput = onforminput;
-	}
-
-	/**
-	 * Gets the onhashchange.
-	 *
-	 * @return the onhashchange
-	 */
-	public Function getOnhashchange() {
-		return onhashchange;
-	}
-
-	/**
-	 * Sets the onhashchange.
-	 *
-	 * @param onhashchange
-	 *            the new onhashchange
-	 */
-	public void setOnhashchange(Function onhashchange) {
-		this.onhashchange = onhashchange;
-	}
-
-	/**
-	 * Gets the oninput.
-	 *
-	 * @return the oninput
-	 */
-	public Function getOninput() {
-		return oninput;
-	}
-
-	/**
-	 * Sets the oninput.
-	 *
-	 * @param oninput
-	 *            the new oninput
-	 */
-	public void setOninput(Function oninput) {
-		this.oninput = oninput;
-	}
-
-	/**
-	 * Gets the oninvalid.
-	 *
-	 * @return the oninvalid
-	 */
-	public Function getOninvalid() {
-		return oninvalid;
-	}
-
-	/**
-	 * Sets the oninvalid.
-	 *
-	 * @param oninvalid
-	 *            the new oninvalid
-	 */
-	public void setOninvalid(Function oninvalid) {
-		this.oninvalid = oninvalid;
 	}
 
 	/**
@@ -2445,7 +1469,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 * @return the onkeydown
 	 */
 	public Function getOnkeydown() {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			return ((HTMLDocumentImpl) doc).getOnkeydown();
 		} else {
@@ -2460,7 +1484,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 *            the new onkeydown
 	 */
 	public void setOnkeydown(Function onkeydown) {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			((HTMLDocumentImpl) doc).setOnkeydown(onkeydown);
 		}
@@ -2472,7 +1496,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 * @return the onkeypress
 	 */
 	public Function getOnkeypress() {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			return ((HTMLDocumentImpl) doc).getOnkeypress();
 		} else {
@@ -2487,7 +1511,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 *            the new onkeypress
 	 */
 	public void setOnkeypress(Function onkeypress) {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			((HTMLDocumentImpl) doc).setOnkeypress(onkeypress);
 		}
@@ -2499,7 +1523,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 * @return the onkeyup
 	 */
 	public Function getOnkeyup() {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			return ((HTMLDocumentImpl) doc).getOnkeyup();
 		} else {
@@ -2514,86 +1538,10 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 *            the new onkeyup
 	 */
 	public void setOnkeyup(Function onkeyup) {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			((HTMLDocumentImpl) doc).setOnkeyup(onkeyup);
 		}
-	}
-
-	/**
-	 * Gets the onloadeddata.
-	 *
-	 * @return the onloadeddata
-	 */
-	public Function getOnloadeddata() {
-		return onloadeddata;
-	}
-
-	/**
-	 * Sets the onloadeddata.
-	 *
-	 * @param onloadeddata
-	 *            the new onloadeddata
-	 */
-	public void setOnloadeddata(Function onloadeddata) {
-		this.onloadeddata = onloadeddata;
-	}
-
-	/**
-	 * Gets the onloadedmetadata.
-	 *
-	 * @return the onloadedmetadata
-	 */
-	public Function getOnloadedmetadata() {
-		return onloadedmetadata;
-	}
-
-	/**
-	 * Sets the onloadedmetadata.
-	 *
-	 * @param onloadedmetadata
-	 *            the new onloadedmetadata
-	 */
-	public void setOnloadedmetadata(Function onloadedmetadata) {
-		this.onloadedmetadata = onloadedmetadata;
-	}
-
-	/**
-	 * Gets the onloadstart.
-	 *
-	 * @return the onloadstart
-	 */
-	public Function getOnloadstart() {
-		return onloadstart;
-	}
-
-	/**
-	 * Sets the onloadstart.
-	 *
-	 * @param onloadstart
-	 *            the new onloadstart
-	 */
-	public void setOnloadstart(Function onloadstart) {
-		this.onloadstart = onloadstart;
-	}
-
-	/**
-	 * Gets the onmessage.
-	 *
-	 * @return the onmessage
-	 */
-	public Function getOnmessage() {
-		return onmessage;
-	}
-
-	/**
-	 * Sets the onmessage.
-	 *
-	 * @param onmessage
-	 *            the new onmessage
-	 */
-	public void setOnmessage(Function onmessage) {
-		this.onmessage = onmessage;
 	}
 
 	/**
@@ -2602,7 +1550,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 * @return the onmousedown
 	 */
 	public Function getOnmousedown() {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			return ((HTMLDocumentImpl) doc).getOnmousedown();
 		} else {
@@ -2617,48 +1565,10 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 *            the new onmousedown
 	 */
 	public void setOnmousedown(Function onmousedown) {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			((HTMLDocumentImpl) doc).setOnmousedown(onmousedown);
 		}
-	}
-
-	/**
-	 * Gets the onmousemove.
-	 *
-	 * @return the onmousemove
-	 */
-	public Function getOnmousemove() {
-		return onmousemove;
-	}
-
-	/**
-	 * Sets the onmousemove.
-	 *
-	 * @param onmousemove
-	 *            the new onmousemove
-	 */
-	public void setOnmousemove(Function onmousemove) {
-		this.onmousemove = onmousemove;
-	}
-
-	/**
-	 * Gets the onmouseout.
-	 *
-	 * @return the onmouseout
-	 */
-	public Function getOnmouseout() {
-		return onmouseout;
-	}
-
-	/**
-	 * Sets the onmouseout.
-	 *
-	 * @param onmouseout
-	 *            the new onmouseout
-	 */
-	public void setOnmouseout(Function onmouseout) {
-		this.onmouseout = onmouseout;
 	}
 
 	/**
@@ -2667,7 +1577,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 * @return the onmouseover
 	 */
 	public Function getOnmouseover() {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			return ((HTMLDocumentImpl) doc).getOnmouseover();
 		} else {
@@ -2682,7 +1592,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 *            the new onmouseover
 	 */
 	public void setOnmouseover(Function onmouseover) {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			((HTMLDocumentImpl) doc).setOnmouseover(onmouseover);
 		}
@@ -2694,7 +1604,7 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 * @return the onmouseup
 	 */
 	public Function getOnmouseup() {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			return ((HTMLDocumentImpl) doc).getOnmouseup();
 		} else {
@@ -2709,523 +1619,10 @@ public class Window extends AbstractScriptableDelegate implements AbstractView {
 	 *            the new onmouseup
 	 */
 	public void setOnmouseup(Function onmouseup) {
-		Document doc = this.document;
+		Document doc = this.getWindowDocument();
 		if (doc instanceof HTMLDocumentImpl) {
 			((HTMLDocumentImpl) doc).setOnmouseup(onmouseup);
 		}
-	}
-
-	/**
-	 * Gets the onmousewheel.
-	 *
-	 * @return the onmousewheel
-	 */
-	public Function getOnmousewheel() {
-		return onmousewheel;
-	}
-
-	/**
-	 * Sets the onmousewheel.
-	 *
-	 * @param onmousewheel
-	 *            the new onmousewheel
-	 */
-	public void setOnmousewheel(Function onmousewheel) {
-		this.onmousewheel = onmousewheel;
-	}
-
-	/**
-	 * Gets the onoffline.
-	 *
-	 * @return the onoffline
-	 */
-	public Function getOnoffline() {
-		return onoffline;
-	}
-
-	/**
-	 * Sets the onoffline.
-	 *
-	 * @param onoffline
-	 *            the new onoffline
-	 */
-	public void setOnoffline(Function onoffline) {
-		this.onoffline = onoffline;
-	}
-
-	/**
-	 * Gets the ononline.
-	 *
-	 * @return the ononline
-	 */
-	public Function getOnonline() {
-		return ononline;
-	}
-
-	/**
-	 * Sets the ononline.
-	 *
-	 * @param ononline
-	 *            the new ononline
-	 */
-	public void setOnonline(Function ononline) {
-		this.ononline = ononline;
-	}
-
-	/**
-	 * Gets the onpause.
-	 *
-	 * @return the onpause
-	 */
-	public Function getOnpause() {
-		return onpause;
-	}
-
-	/**
-	 * Sets the onpause.
-	 *
-	 * @param onpause
-	 *            the new onpause
-	 */
-	public void setOnpause(Function onpause) {
-		this.onpause = onpause;
-	}
-
-	/**
-	 * Gets the onplay.
-	 *
-	 * @return the onplay
-	 */
-	public Function getOnplay() {
-		return onplay;
-	}
-
-	/**
-	 * Sets the onplay.
-	 *
-	 * @param onplay
-	 *            the new onplay
-	 */
-	public void setOnplay(Function onplay) {
-		this.onplay = onplay;
-	}
-
-	/**
-	 * Gets the onplaying.
-	 *
-	 * @return the onplaying
-	 */
-	public Function getOnplaying() {
-		return onplaying;
-	}
-
-	/**
-	 * Sets the onplaying.
-	 *
-	 * @param onplaying
-	 *            the new onplaying
-	 */
-	public void setOnplaying(Function onplaying) {
-		this.onplaying = onplaying;
-	}
-
-	/**
-	 * Gets the onpagehide.
-	 *
-	 * @return the onpagehide
-	 */
-	public Function getOnpagehide() {
-		return onpagehide;
-	}
-
-	/**
-	 * Sets the onpagehide.
-	 *
-	 * @param onpagehide
-	 *            the new onpagehide
-	 */
-	public void setOnpagehide(Function onpagehide) {
-		this.onpagehide = onpagehide;
-	}
-
-	/**
-	 * Gets the onpageshow.
-	 *
-	 * @return the onpageshow
-	 */
-	public Function getOnpageshow() {
-		return onpageshow;
-	}
-
-	/**
-	 * Sets the onpageshow.
-	 *
-	 * @param onpageshow
-	 *            the new onpageshow
-	 */
-	public void setOnpageshow(Function onpageshow) {
-		this.onpageshow = onpageshow;
-	}
-
-	/**
-	 * Gets the onpopstate.
-	 *
-	 * @return the onpopstate
-	 */
-	public Function getOnpopstate() {
-		return onpopstate;
-	}
-
-	/**
-	 * Sets the onpopstate.
-	 *
-	 * @param onpopstate
-	 *            the new onpopstate
-	 */
-	public void setOnpopstate(Function onpopstate) {
-		this.onpopstate = onpopstate;
-	}
-
-	/**
-	 * Gets the onprogress.
-	 *
-	 * @return the onprogress
-	 */
-	public Function getOnprogress() {
-		return onprogress;
-	}
-
-	/**
-	 * Sets the onprogress.
-	 *
-	 * @param onprogress
-	 *            the new onprogress
-	 */
-	public void setOnprogress(Function onprogress) {
-		this.onprogress = onprogress;
-	}
-
-	/**
-	 * Gets the onratechange.
-	 *
-	 * @return the onratechange
-	 */
-	public Function getOnratechange() {
-		return onratechange;
-	}
-
-	/**
-	 * Sets the onratechange.
-	 *
-	 * @param onratechange
-	 *            the new onratechange
-	 */
-	public void setOnratechange(Function onratechange) {
-		this.onratechange = onratechange;
-	}
-
-	/**
-	 * Gets the onreadystatechange.
-	 *
-	 * @return the onreadystatechange
-	 */
-	public Function getOnreadystatechange() {
-		return onreadystatechange;
-	}
-
-	/**
-	 * Sets the onreadystatechange.
-	 *
-	 * @param onreadystatechange
-	 *            the new onreadystatechange
-	 */
-	public void setOnreadystatechange(Function onreadystatechange) {
-		this.onreadystatechange = onreadystatechange;
-	}
-
-	/**
-	 * Gets the onredo.
-	 *
-	 * @return the onredo
-	 */
-	public Function getOnredo() {
-		return onredo;
-	}
-
-	/**
-	 * Sets the onredo.
-	 *
-	 * @param onredo
-	 *            the new onredo
-	 */
-	public void setOnredo(Function onredo) {
-		this.onredo = onredo;
-	}
-
-	/**
-	 * Gets the onresize.
-	 *
-	 * @return the onresize
-	 */
-	public Function getOnresize() {
-		return onresize;
-	}
-
-	/**
-	 * Sets the onresize.
-	 *
-	 * @param onresize
-	 *            the new onresize
-	 */
-	public void setOnresize(Function onresize) {
-		this.onresize = onresize;
-	}
-
-	/**
-	 * Gets the onscroll.
-	 *
-	 * @return the onscroll
-	 */
-	public Function getOnscroll() {
-		return onscroll;
-	}
-
-	/**
-	 * Sets the onscroll.
-	 *
-	 * @param onscroll
-	 *            the new onscroll
-	 */
-	public void setOnscroll(Function onscroll) {
-		this.onscroll = onscroll;
-	}
-
-	/**
-	 * Gets the onseeked.
-	 *
-	 * @return the onseeked
-	 */
-	public Function getOnseeked() {
-		return onseeked;
-	}
-
-	/**
-	 * Sets the onseeked.
-	 *
-	 * @param onseeked
-	 *            the new onseeked
-	 */
-	public void setOnseeked(Function onseeked) {
-		this.onseeked = onseeked;
-	}
-
-	/**
-	 * Gets the onseeking.
-	 *
-	 * @return the onseeking
-	 */
-	public Function getOnseeking() {
-		return onseeking;
-	}
-
-	/**
-	 * Sets the onseeking.
-	 *
-	 * @param onseeking
-	 *            the new onseeking
-	 */
-	public void setOnseeking(Function onseeking) {
-		this.onseeking = onseeking;
-	}
-
-	/**
-	 * Gets the onselect.
-	 *
-	 * @return the onselect
-	 */
-	public Function getOnselect() {
-		return onselect;
-	}
-
-	/**
-	 * Sets the onselect.
-	 *
-	 * @param onselect
-	 *            the new onselect
-	 */
-	public void setOnselect(Function onselect) {
-		this.onselect = onselect;
-	}
-
-	/**
-	 * Gets the onshow.
-	 *
-	 * @return the onshow
-	 */
-	public Function getOnshow() {
-		return onshow;
-	}
-
-	/**
-	 * Sets the onshow.
-	 *
-	 * @param onshow
-	 *            the new onshow
-	 */
-	public void setOnshow(Function onshow) {
-		this.onshow = onshow;
-	}
-
-	/**
-	 * Gets the onstalled.
-	 *
-	 * @return the onstalled
-	 */
-	public Function getOnstalled() {
-		return onstalled;
-	}
-
-	/**
-	 * Sets the onstalled.
-	 *
-	 * @param onstalled
-	 *            the new onstalled
-	 */
-	public void setOnstalled(Function onstalled) {
-		this.onstalled = onstalled;
-	}
-
-	/**
-	 * Gets the onstorage.
-	 *
-	 * @return the onstorage
-	 */
-	public Function getOnstorage() {
-		return onstorage;
-	}
-
-	/**
-	 * Sets the onstorage.
-	 *
-	 * @param onstorage
-	 *            the new onstorage
-	 */
-	public void setOnstorage(Function onstorage) {
-		this.onstorage = onstorage;
-	}
-
-	/**
-	 * Gets the onsubmit.
-	 *
-	 * @return the onsubmit
-	 */
-	public Function getOnsubmit() {
-		return onsubmit;
-	}
-
-	/**
-	 * Sets the onsubmit.
-	 *
-	 * @param onsubmit
-	 *            the new onsubmit
-	 */
-	public void setOnsubmit(Function onsubmit) {
-		this.onsubmit = onsubmit;
-	}
-
-	/**
-	 * Gets the onsuspend.
-	 *
-	 * @return the onsuspend
-	 */
-	public Function getOnsuspend() {
-		return onsuspend;
-	}
-
-	/**
-	 * Sets the onsuspend.
-	 *
-	 * @param onsuspend
-	 *            the new onsuspend
-	 */
-	public void setOnsuspend(Function onsuspend) {
-		this.onsuspend = onsuspend;
-	}
-
-	/**
-	 * Gets the ontimeupdate.
-	 *
-	 * @return the ontimeupdate
-	 */
-	public Function getOntimeupdate() {
-		return ontimeupdate;
-	}
-
-	/**
-	 * Sets the ontimeupdate.
-	 *
-	 * @param ontimeupdate
-	 *            the new ontimeupdate
-	 */
-	public void setOntimeupdate(Function ontimeupdate) {
-		this.ontimeupdate = ontimeupdate;
-	}
-
-	/**
-	 * Gets the onundo.
-	 *
-	 * @return the onundo
-	 */
-	public Function getOnundo() {
-		return onundo;
-	}
-
-	/**
-	 * Sets the onundo.
-	 *
-	 * @param onundo
-	 *            the new onundo
-	 */
-	public void setOnundo(Function onundo) {
-		this.onundo = onundo;
-	}
-
-	/**
-	 * Gets the onvolumechange.
-	 *
-	 * @return the onvolumechange
-	 */
-	public Function getOnvolumechange() {
-		return onvolumechange;
-	}
-
-	/**
-	 * Sets the onvolumechange.
-	 *
-	 * @param onvolumechange
-	 *            the new onvolumechange
-	 */
-	public void setOnvolumechange(Function onvolumechange) {
-		this.onvolumechange = onvolumechange;
-	}
-
-	/**
-	 * Gets the onwaiting.
-	 *
-	 * @return the onwaiting
-	 */
-	public Function getOnwaiting() {
-		return onwaiting;
-	}
-
-	/**
-	 * Sets the onwaiting.
-	 *
-	 * @param onwaiting
-	 *            the new onwaiting
-	 */
-	public void setOnwaiting(Function onwaiting) {
-		this.onwaiting = onwaiting;
 	}
 
 	/**
